feat: Create Home Page layout
