--- lib/plantar_web/live/page_live.ex
@@ -1,8 +1,0 @@
-defmodule PlantarWeb.PageLive do
-  use PlantarWeb, :live_view
-
-  @impl true
-  def mount(_params, _session, socket) do
-    {:ok,socket}
-  end
-end
