--- lib/plantar_web/live/page_live.html.leex
@@ -1,3 +1,0 @@
-<section class="container mx-auto">
-  <h1 class="text-lg text-orange-500"><%= gettext "Welcome to %{name}!", name: "Phoenix" %></h1>
-</section>
