--- lib/plantar_web/templates/crop/index.html.leex
@@ -1,53 +1,0 @@
-<h1>Listing Crops</h1>
-
-<%= if @live_action in [:new, :edit] do %>
-  <%= live_modal @socket, PlantarWeb.CropLive.FormComponent,
-    id: @crop.id || :new,
-    title: @page_title,
-    user_id: @crop.user_id || @current_user.id,
-    action: @live_action,
-    crop: @crop,
-    return_to: Routes.crop_index_path(@socket, :index) %>
-<% end %>
-
-<table>
-  <thead>
-    <tr>
-      <th>Name</th>
-      <th>Alternatives names</th>
-      <th>Binomial name</th>
-      <th>Description</th>
-      <th>Height</th>
-      <th>Row spacing</th>
-      <th>Days of maturity</th>
-      <th>Sun requirements</th>
-
-      <th></th>
-    </tr>
-  </thead>
-  <tbody id="crops">
-    <%= for crop <- @crops do %>
-      <tr id="crop-<%= crop.id %>">
-        <td><%= crop.name %></td>
-        <td><%= crop.alternatives_names %></td>
-        <td><%= crop.binomial_name %></td>
-        <td><%= crop.description %></td>
-        <td><%= crop.height %></td>
-        <td><%= crop.row_spacing %></td>
-        <td><%= crop.days_of_maturity %></td>
-        <td><%= crop.sun_requirements %></td>
-
-        <td>
-          <span><%= live_redirect "Show", to: Routes.crop_show_path(@socket, :show, crop) %></span>
-          <%= if @current_user && @current_user.admin do %>
-            <span><%= live_patch "Edit", to: Routes.crop_index_path(@socket, :edit, crop) %></span>
-            <span><%= link "Delete", to: "#", phx_click: "delete", phx_value_id: crop.id, data: [confirm: "Are you sure?"] %></span>
-          <% end %>
-        </td>
-      </tr>
-    <% end %>
-  </tbody>
-</table>
-<%= if @current_user && @current_user.admin  do %>
-  <span><%= live_patch "New Crop", to: Routes.crop_index_path(@socket, :new) %></span>
-<% end %>
