+++ lib/plantar_web/templates/home/index.html.leex
@@ -1,0 +1,26 @@
+<div class="w-10/12 mx-auto">
+  <div class="w-full md:w-2/3 flex flex-col mt-10 mb-16 mx-auto text-center">
+    <h1 class="title-font sm:text-4xl text-3xl mb-4 font-medium text-gray-900">O que você quer plantar?</p>
+    <div class="flex w-full justify-center mt-4">
+      <input class="border-0 bg-gray-100 rounded mr-4 border border-gray-400 focus:outline-none focus:border-indigo-500 text-base px-4 py-2 lg:w-full xl:w-1/2 w-2/4 md:w-full" placeholder="Tomate" type="text">
+    </div>
+  </div>
+  <div class="w-full flex flex-wrap -m-4">
+    <%= for crop <- @crops do%>
+      <div class="p-4 md:w-1/3">
+        <%= live_redirect to: Routes.crop_show_path(@socket, :show, crop) do %>
+          <div class="h-full border-2 border-gray-200 rounded-lg overflow-hidden">
+            <img class="lg:h-48 md:h-36 w-full object-cover object-center" src="https://saude.abril.com.br/wp-content/uploads/2016/12/mandioca.jpg?quality=85&amp;strip=info&amp;w=680" alt="blog">
+            <div class="p-6">
+              <h2 class="tracking-widest text-sm title-font font-medium text-gray-600 mb-1"><%= crop.alternatives_names %> </h2>
+              <h1 class="title-font text-xl font-medium text-gray-900 mb-3"><%= crop.name %></h1>
+            </div>
+          </div>
+        <% end %>
+      </div>
+    <% end %>
+  </div>
+  <%= if @current_user && @current_user.admin  do %>
+    <span><%= live_patch "New Crop", to: Routes.crop_index_path(@socket, :new) %></span>
+  <% end %>
+</div>
