+++ lib/plantar_web/views/home_view.ex
@@ -1,0 +1,3 @@
+defmodule PlantarWeb.HomeView do
+  use PlantarWeb, :view
+end
